requirements.txt
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
ebaysdk==2.2.0
idna==2.10
lxml==4.6.3
pymongo==3.11.3
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.4

AsynceBay~=1.0.0
psycopg2~=2.9.3
python-dateutil~=2.8.2